+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ## Maintenance
 
+- Added a check for docs in the `CI` ([#59](https://github.com/Saransh-cpp/OCRed/pull/58))
+- Fixed failing doc deployment ([#59](https://github.com/Saransh-cpp/OCRed/pull/58))
 - Fixed building from archive (tarballs) ([#56](https://github.com/Saransh-cpp/OCRed/pull/56))
 
 # [v0.1.2](https://github.com/Saransh-cpp/OCRed/tree/v0.1.2)